requirements.txt
scikit-learn==1.5.1
matplotlib==3.9.2
nltk==3.9.1
numpy>=1.19.5
tensorflow==2.15.0
tf-keras==2.15.0
transformers==4.37.0